Separate PersonalClient pages
