+++ src/screens/PersonalDataAddress/components/newAddress/Form/index.tsx
@@ -4,10 +4,11 @@ import { Button, Card, Grid, TextField, Typography } from '@mui/material';
 //import address from '../../../../tests/mock/Address/index.json';
 import { Address } from '../../../../../lib/interfaces/Address';
 //import { Link } from 'react-router-dom';
+import { v4 as uuidv4 } from 'uuid';
 const FormNewAddress: React.FC = () => {
   const style = useCustomStyles();
   const [userDatas, setUserDatas] = useState<Address>({
-    id: '',
+    id: uuidv4(),
     userId: '',
     apelido: '',
     cep: '',
@@ -47,7 +48,7 @@ const FormNewAddress: React.FC = () => {
           </Typography>
         </Grid>
         <Grid item md={12}>
-          <Card>
+          <Card sx={{ p: '2em' }}>
             <TextField
               fullWidth
               label="APELIDO"
@@ -132,7 +133,7 @@ const FormNewAddress: React.FC = () => {
               onClick={handleSubmitClick}
               sx={style.Button}
             >
-              Salvar
+              Cadastrar
             </Button>
           </Card>
         </Grid>
