+++ src/screens/PersonalDataCard/components/PersonalCardsList/style.ts
