+++ src/screens/PersonalDataCard/components/personalCards/style.ts
