+++ src/screens/PersonalDataClient/index.tsx
@@ -3,21 +3,24 @@ import { Container, Grid } from '@mui/material';
 import { useCustomStyles } from './style';
 import SmallFooter from '../../lib/components/Footer/smallFooter';
 import Navbar from '../../lib/components/Navbar/Navbar';
-import { useState } from 'react';
+//import { useState } from 'react';
 import PersonalInfo from './components/personalData';
-import PersonalWishlist from './components/personalWishlist';
-import PersonalCardsList from './components/PersonalCardsList';
-import PersonalAddressList from './components/PersonalAddressList';
-import SideMenu from './components/sideMenu';
+//import PersonalWishlist from './components/personalWishlist';
+//import PersonalCardsList from './components/PersonalCardsList';
+//import PersonalAddressList from './components/PersonalAddressList';
+import SideMenu from '../../lib/components/SideMenu';
+//import FormNewAddress from './components/newAddress/Form';
 export default function PersonalDataClient() {
   const style = useCustomStyles();
-  const [showPersonalInfo, setShowPersonalInfo] = useState(true);
+  /*const [showPersonalInfo, setShowPersonalInfo] = useState(true);
   const [showPersonalAddress, setShowPersonalAddress] = useState(false);
+  const [showFormNewAddress, setShowFormNewAddress] = useState(false);
   const [showPersonalCards, setShowPersonalCards] = useState(false);
   const [showPersonalWishlist, setShowPersonalWishlist] = useState(false);
   const [selectedItem, setSelectedItem] = useState('dados-pessoais');
   const handleItemClick = (key: string) => {
     setSelectedItem(key);
+    setShowFormNewAddress(false);
     if (key === 'dados-pessoais') {
       setShowPersonalInfo(true);
       setShowPersonalAddress(false);
@@ -40,35 +43,21 @@ export default function PersonalDataClient() {
       setShowPersonalWishlist(true);
     }
   };
-
+  const handleNewAddressClick = () => {
+    setShowFormNewAddress(true);
+    setShowPersonalInfo(false);
+    setShowPersonalAddress(false);
+    setShowPersonalCards(false);
+    setShowPersonalWishlist(false);
+  };*/
   return (
     <>
       <Navbar isHomePage={false} />
       <Grid container sx={style.Content}>
-        <SideMenu
-          selectedItem={selectedItem}
-          handleItemClick={handleItemClick}
-        />
-        {showPersonalInfo && (
-          <Grid container item xs={12} sm={12} md={9} sx={style.ContainerStyle}>
-            <PersonalInfo />
-          </Grid>
-        )}
-        {showPersonalAddress && (
-          <Grid container item xs={12} sm={12} md={9} sx={style.ContainerStyle}>
-            <PersonalAddressList />
-          </Grid>
-        )}
-        {showPersonalCards && (
-          <Grid container item xs={12} sm={12} md={9} sx={style.ContainerStyle}>
-            <PersonalCardsList />
-          </Grid>
-        )}
-        {showPersonalWishlist && (
-          <Grid container item xs={12} sm={12} md={9} sx={style.ContainerStyle}>
-            <PersonalWishlist />
-          </Grid>
-        )}
+        <SideMenu />
+        <Grid container item xs={12} sm={12} md={9} sx={style.ContainerStyle}>
+          <PersonalInfo />
+        </Grid>
       </Grid>
       <Container sx={style.SmallFooterStyle}>
         <SmallFooter />
